# swagger.yaml
openapi: 3.0.3
info:
  title: API Document for Mor Marketplace
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:3001/api
...
paths:
...
  /user/add-user:
    post:
      tags:
        - User
      summary: Add a new user
      description: avatar is image
      operationId: addUser
      requestBody:
        description: Create a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
          application/xml:
            schema:
              $ref: '#/components/schemas/UserCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'          
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: User already exist
        '403':
          description: Username already exist
...
  /category/get-by-id/{id}:
    get:
      tags:
        - Category
      summary: get category by id
      description: id is categoryId
      operationId: getCategoryById
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
  /category/add-category:
    post:
      tags:
        - Category
      summary: Add a new category
      description: Add a new category
      operationId: addCategory
      
      requestBody:
        description: Create a new Category
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
          application/xml:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'   
        '400':
          description: Category already exist
  /category/update/{id}:
    put:
      tags:
        - Category
      summary: Add a new category
      description: Add a new category
      operationId: updateCategory
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        description: Create a new Category
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
          application/xml:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Category'   
        '404':
          description: Category not found
  /category/delete/{id}:
    delete:
      tags:
        - Category
      summary: Delete category
      description: Delete category
      operationId: deleteCategory
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          type: integer
      responses:
          '200':
            description: Successful operation
            content:
              application/json:
                example: "category deleted"
          '400':
            description: Category already exist
...
components:
  schemas:
    User:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        email:
          type: string
        password:
          type: string
        name:
          type: string
        avatar:
          type: string
        verified:
          type: boolean
        role:
          type: string
          enum:
            - USER
            - ADMIN
        createdAt:
          type: string
          format: date-time
    UserCreate:
        type: object
        properties:
          username:
            type: string
          email:
            type: string
          password:
            type: string
          name:
            type: string
          avatar:
            type: string
...
    Category:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
    CategoryCreate:
      type: object
      properties:
        categoryName: 
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT